Unlimited primer length
Fixed a bug that caused failure in PCR simulation when primer's length exceeded the maximum "effective" length. It now works the way it was intended, with the effective (3') end of the primer being used in the search for matches, but the entire primer -- regardless of length -- used to construct the amplicon sequence. This change avoids the need to increase the maximum effective length, as in the previous update.